refactor(ai): extract draft-session + discussion-reply helpers (U3)
